Refactor RsExtractEnumVariantAction
